Fixes 1619: Refactor Fabric's BDD/Behave tests - step 1
