--- bddtests/steps/bdd_test_util.py
@@ -1,44 +1,0 @@
-
-# Copyright IBM Corp. 2016 All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-import os
-import subprocess
-
-
-def cli_call(context, arg_list, expect_success=True):
-    """Executes a CLI command in a subprocess and return the results.
-
-    @param context: the behave context
-    @param arg_list: a list command arguments
-    @param expect_success: use False to return even if an error occurred when executing the command
-    @return: (string, string, int) output message, error message, return code
-    """
-    #arg_list[0] = "update-" + arg_list[0]
-
-    # We need to run the cli command by actually calling the python command
-    # the update-cli.py script has a #!/bin/python as the first line
-    # which calls the system python, not the virtual env python we
-    # setup for running the update-cli
-    p = subprocess.Popen(arg_list, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    output, error = p.communicate()
-    if p.returncode != 0:
-        if output is not None:
-            print("Output:\n" + output)
-        if error is not None:
-            print("Error Message:\n" + error)
-        if expect_success:
-            raise subprocess.CalledProcessError(p.returncode, arg_list, output)
-    return output, error, p.returncode
