+++ bddtests/steps/shared/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+"""
+Used for behave as compatibility layer between different Python versions/implementations.
+"""
